requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
pytest==8.0.0
black==24.1.1
flake8==7.0.0
PyMySQL==1.1.0 